backend/Lip-Reading/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
boto3==1.26.102
botocore==1.29.102
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
# chatGPT @ git+https://github.com/mmabrouk/chatgpt-wrapper@b794c56b18bf3467617c2f82a9b60c1c4932a808
click==8.1.3
comm==0.1.3
dataclasses-json==0.5.7
debugpy==1.6.6
decorator==4.4.2
defusedxml==0.7.1
# dlib==19.24.1
dnspython==2.3.0
email-validator==1.3.1
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
Flask==2.2.3
flatbuffers==23.3.3
fqdn==1.5.1
frozenlist==1.3.3
gast==0.4.0
geojson==2.5.0
google-auth==2.16.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.53.0
h5py==3.8.0
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
iniconfig==2.0.0

ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.7
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
keras==2.12.0
langchain==0.0.126
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
meson==1.0.1
mistune==2.0.5
ml-dtypes==0.0.4
moviepy==1.0.3
multidict==6.0.4
mypy-extensions==1.0.0
names==0.3.0
nbclassic==0.5.4
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.2
openai==0.27.2
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
playwright==1.32.1
pluggy==1.0.0
proglog==0.1.10
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
pyee==9.0.4
Pygments==2.14.0
pyowm==3.3.0
pyperclip==1.8.2
pyreadline3==3.4.1
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.2.2
pytest-base-url==2.0.0
pytest-playwright==0.3.2
python-dateutil==2.8.2
python-frontmatter==1.0.0
python-json-logger==2.0.7
python-slugify==6.1.2
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.3
rsa==4.9
s3transfer==0.6.0
scipy==1.10.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.47
stack-data==0.6.2
tenacity==8.2.2
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
tiktoken==0.3.3
tinycss2==1.2.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.15
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.14.1
yarl==1.8.2
